segregate navBar component
- the index file is getting quite large
